refactor and reorganise for sanity
