+++ src/main/java/org/cancogenvirusseq/singularity/components/Analyses.java
@@ -1,0 +1,64 @@
+/*
+ * Copyright (c) 2021 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ * This program and the accompanying materials are made available under the terms of the GNU Affero General Public License v3.0.
+ * You should have received a copy of the GNU Affero General Public License along with
+ * this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package org.cancogenvirusseq.singularity.components;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.cancogenvirusseq.singularity.components.model.AnalysisDocument;
+import org.cancogenvirusseq.singularity.config.elasticsearch.ElasticsearchProperties;
+import org.cancogenvirusseq.singularity.config.elasticsearch.ReactiveElasticSearchClientConfig;
+import org.elasticsearch.action.search.SearchRequest;
+import org.elasticsearch.index.query.QueryBuilders;
+import org.elasticsearch.search.SearchHit;
+import org.elasticsearch.search.builder.SearchSourceBuilder;
+import org.springframework.stereotype.Component;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class Analyses {
+  private final ElasticsearchProperties elasticsearchProperties;
+  private final ReactiveElasticSearchClientConfig reactiveElasticSearchClientConfig;
+  private final ObjectMapper objectMapper;
+
+  public Flux<AnalysisDocument> getAllAnalysisDocuments() {
+    return Mono.just(
+            new SearchSourceBuilder()
+                .query(QueryBuilders.matchAllQuery())
+                .fetchSource(AnalysisDocument.getEsIncludeFields(), null))
+        .flatMapMany(
+            source ->
+                reactiveElasticSearchClientConfig
+                    .reactiveElasticsearchClient()
+                    .scroll(
+                        new SearchRequest()
+                            .indices(elasticsearchProperties.getFileCentricIndex())
+                            .source(source)))
+        .map(this::hitMapToAnalysisDocument);
+  }
+
+  @SneakyThrows
+  private AnalysisDocument hitMapToAnalysisDocument(SearchHit hit) {
+    return objectMapper.readValue(hit.getSourceAsString(), AnalysisDocument.class);
+  }
+}
