+++ src/main/java/org/cancogenvirusseq/singularity/components/Files.java
@@ -18,37 +18,27 @@
 
 package org.cancogenvirusseq.singularity.components;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
+import static org.cancogenvirusseq.singularity.components.FileArchiveUtils.batchedDownloadPairsToFileArchiveWithInstant;
+
 import java.time.Duration;
 import java.time.Instant;
 import java.util.concurrent.atomic.AtomicReference;
 import javax.annotation.PostConstruct;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.cancogenvirusseq.singularity.components.events.EventEmitter;
-import org.cancogenvirusseq.singularity.components.model.AnalysisDocument;
-import org.cancogenvirusseq.singularity.config.elasticsearch.ElasticsearchProperties;
-import org.cancogenvirusseq.singularity.config.elasticsearch.ReactiveElasticSearchClientConfig;
-import org.elasticsearch.action.search.SearchRequest;
-import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.search.SearchHit;
-import org.elasticsearch.search.builder.SearchSourceBuilder;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import reactor.core.Disposable;
 import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
 
 @Slf4j
 @Component
 @RequiredArgsConstructor
 public class Files {
-  private final ElasticsearchProperties elasticsearchProperties;
-  private final ReactiveElasticSearchClientConfig reactiveElasticSearchClientConfig;
-  private final ObjectMapper objectMapper;
   private final EventEmitter eventEmitter;
+  private final Analyses analyses;
   private final Download download;
 
   @Value("${files.triggerUpdateDelaySeconds}")
@@ -103,29 +93,10 @@ public class Files {
   }
 
   private Flux<String> downloadAndSave(Instant instant) {
-    return getAllAnalysisDocuments()
-        .transform(download.downloadAndArchiveFunctionWithInstant(instant))
+    return analyses
+        .getAllAnalysisDocuments()
+        .transform(download::downloadBatchedPairs)
+        .transform(batchedDownloadPairsToFileArchiveWithInstant(instant))
         .log("Files::downloadAndSave");
-  }
-
-  private Flux<AnalysisDocument> getAllAnalysisDocuments() {
-    return Mono.just(
-            new SearchSourceBuilder()
-                .query(QueryBuilders.matchAllQuery())
-                .fetchSource(AnalysisDocument.getEsIncludeFields(), null))
-        .flatMapMany(
-            source ->
-                reactiveElasticSearchClientConfig
-                    .reactiveElasticsearchClient()
-                    .scroll(
-                        new SearchRequest()
-                            .indices(elasticsearchProperties.getFileCentricIndex())
-                            .source(source)))
-        .map(this::hitMapToAnalysisDocument);
-  }
-
-  @SneakyThrows
-  private AnalysisDocument hitMapToAnalysisDocument(SearchHit hit) {
-    return objectMapper.readValue(hit.getSourceAsString(), AnalysisDocument.class);
   }
 }
